+++ go.mod
@@ -1,3 +1,11 @@
 module github.com/vantmet/trackmyrun
 
 go 1.17
+
+require (
+	github.com/aws/aws-cdk-go/awscdk/v2 v2.14.0
+	github.com/aws/constructs-go/constructs/v10 v10.0.9
+	github.com/aws/jsii-runtime-go v1.54.0
+)
+
+require github.com/Masterminds/semver/v3 v3.1.1 // indirect
